Add files via upload
The Sudoku solver project utilizes Python with the tkinter package for GUI creation and pillow for image support. It offers a user-friendly interface where users can input Sudoku puzzles into an adjustable grid. The "Solve" button initiates the solver algorithm, attempting to find a solution. The project is designed for ease of use, providing real-time feedback on puzzle solutions or indicating if no solution exists. This tool is useful for individuals seeking an interactive and visually intuitive way to solve Sudoku puzzles.